Record now takes driver in intialization

+++ xplinter/drivers/csv_driver.py
@@ -1,9 +1,15 @@
 from xplinter import Record
 import os
-from .driver import Driver
+from xplinter.driver import Driver
 
 class Csv_driver(Driver):
-    def __init__(self, record: Record, directory: str, reset: bool = False):
+    def __init__(self, directory: str, reset: bool = False):
+        self.tables: dict = {}
+        self._directory = directory
+        self._open: bool = False
+        if reset:
+            raise NotImplementedError
+    def set_record(self, record: Record):
         self.tables: dict = {}
         for entity_name, entity in record.entity_dict.items():
             if entity_name.startswith('*'):
@@ -11,16 +17,19 @@ class Csv_driver(Driver):
             self.tables[entity_name] = [field.name for field in entity.field_list if not field.name.startswith('*')]
         for view_name, view in record.view_dict.items():
             self.tables[view_name] = view.columns[:]
-        self._directory = directory
-        self.is_compatible()
-        self._record = record
+        self._record: Record = record
     def open(self):
+        if not hasattr(self, '_record'): raise RuntimeError('CSV driver cannot open unless record is set')
+        if self._open: raise RuntimeError('CSV driver already open')
+        self.is_compatible()
         for table_name, columns in self.tables.items():
             filename = os.path.join(self._directory, table_name + '.csv')
             if not os.path.exists(filename):
                 with open(filename, 'w') as f:
                     f.write(','.join(columns) + '\n')
+        self._open = True
     def is_compatible(self) -> bool:
+        if not self._record: raise RuntimeError('CSV driver cannot check compatibility unless record is set')
         if not os.path.isdir(self._directory):
             raise RuntimeError(f'Directory `{self._directory}` does not exist')
         for table_name, columns in self.tables.items():
@@ -40,6 +49,7 @@ class Csv_driver(Driver):
             with open(filename, 'w') as f:
                 f.write(','.join(columns) + '\n')
     def write(self):
+        if not self._open: raise RuntimeError('CSV driver `write` attempted while driver not open')
         for entity_name, entity in self._record.entity_dict.items():
             if entity_name.startswith('*'):
                 continue
@@ -49,4 +59,5 @@ class Csv_driver(Driver):
             df = view.to_dataframe()
             df.to_csv(os.path.join(self._directory, view_name + '.csv'), mode='a', index=False, header=False)
     def close(self):
-        pass # Nothing to do for CSV writer
+        if not self._open: raise RuntimeError('CSV driver `close` attempted while driver not open')
+        self._open = False # Nothing else to do for CSV writer
